Add files via upload
Hello World. I made a simple Machine Learning project using the Iris Dataset from Kaggle. The data set consists of 50 samples from each of three species of Iris (Iris setosa, Iris virginica and Iris versicolor). Four features were measured from each sample: the length and the width of the sepals and petals, in centimeters. I have implemented the K-means Algorithm in this project. K-means Algorithm is a clustering algorithm which allows us to segregate and plot the data in the form of different graphs according to its similar features.